backend/requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.20.0
websockets>=12.0
peewee>=3.18.0
python-multipart>=0.0.6
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0